# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Ptium API
  version: 0.1.0
  summary: AI presentation generation, administration, and MCP API
  description: |
    Ptium exposes owner-scoped presentation generation, profile and API-key
    management, administrator operations, and a stateless MCP endpoint.

    JSON REST responses use an envelope containing either `data` or `error`
    plus `requestId`. List responses additionally contain `meta`. The `/mcp`
    endpoint uses raw JSON-RPC 2.0 messages and does not use the REST envelope.

    JSON request bodies are limited to 2 MiB, reject unknown top-level fields,
    and must contain exactly one complete JSON value. UUID path parameters are
    validated before their operation handler runs.

    OIDC access tokens and Ptium API keys are both accepted as Bearer
    credentials. Ptium API keys may alternatively be sent in `X-API-Key`.
    Development authentication is disabled by default and uses
    `X-Ptium-Dev-Secret` unless the server is configured with another header.

    The `x-ptium-api-key-scopes` extension on an operation names the scope a
    Ptium API key must grant. Interactive OIDC and development sessions are
    not restricted by API-key scopes. MCP API-key calls always require
    `mcp:use`; each presentation tool additionally requires its documented
    `presentations:read` or `presentations:write` scope.
servers:
  - url: /
    description: Same origin as the Ptium web application
...
security:
  - BearerAuth: []
  - APIKeyHeader: []
  - DevSecret: []
paths:
...
  /api/v1/presentations/{id}:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
      - $ref: '#/components/parameters/PresentationID'
    get:
      operationId: getPresentation
      summary: Get a presentation and its slides
      tags: [Presentations]
      x-ptium-api-key-scopes: [presentations:read]
      responses:
        '200':
          description: Presentation owned by the current user.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      operationId: replacePresentation
      summary: Update presentation metadata and optionally replace slides
      description: |
        Metadata fields are merged with the current presentation. When `slides`
        is present, the entire slide sequence is atomically replaced; array
        order determines the saved `position`.
      tags: [Presentations]
      x-ptium-api-key-scopes: [presentations:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresentationUpdate'
      responses:
        '200':
          description: Updated presentation including its current slides.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updatePresentation
      summary: Partially update presentation metadata or replace slides
      description: When `slides` is present, the entire sequence is replaced; array order is authoritative.
      tags: [Presentations]
      x-ptium-api-key-scopes: [presentations:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresentationUpdate'
      responses:
        '200':
          description: Updated presentation including its current slides.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deletePresentation
      summary: Move a presentation to the recycle bin
      tags: [Presentations]
      x-ptium-api-key-scopes: [presentations:write]
      responses:
        '204':
          description: Presentation moved to the recycle bin. The response has no body.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/templates:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    get:
      operationId: listTemplates
      summary: List usable PowerPoint templates
      description: |
        Returns the caller's own uploads together with every template shared
        with the organisation and every built-in design. List records omit the
        full manifest; read a single template to obtain its layouts.
      tags: [Templates]
      x-ptium-api-key-scopes: [templates:read]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Templates ordered with uploads first, then built-in designs.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createTemplate
      summary: Upload a PowerPoint template
      description: |
        Accepts a `.pptx` or `.potx` package as either a multipart form or a raw
        request body with metadata in the query string. The package is analyzed
        on upload: its masters, layouts, theme and placeholder geometry are
        catalogued, and the original file is stored unchanged for export.
      tags: [Templates]
      x-ptium-api-key-scopes: [templates:write]
      parameters:
        - name: name
          in: query
          required: false
          schema: { type: string, maxLength: 120 }
        - name: description
          in: query
          required: false
          schema: { type: string, maxLength: 1000 }
        - name: filename
          in: query
          required: false
          description: Required when the body is raw so the extension can be validated.
          schema: { type: string, maxLength: 200 }
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [private, shared]
            default: private
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
                  maxLength: 120
                description:
                  type: string
                  maxLength: 1000
                scope:
                  type: string
                  enum: [private, shared]
          application/vnd.openxmlformats-officedocument.presentationml.presentation:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Template stored and analyzed.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Template uploads are disabled by the administrator.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          description: The upload exceeds the configured template size limit.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/templates/{id}:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
      - $ref: '#/components/parameters/TemplateID'
    get:
      operationId: getTemplate
      summary: Read a template and its layout catalogue
      tags: [Templates]
      x-ptium-api-key-scopes: [templates:read]
      responses:
        '200':
          description: Template including the full manifest.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateTemplate
      summary: Rename a template or change its sharing scope
      description: Built-in designs cannot be modified.
      tags: [Templates]
      x-ptium-api-key-scopes: [templates:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateUpdate'
      responses:
        '200':
          description: Template updated.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteTemplate
      summary: Delete an uploaded template
      description: |
        Presentations that referenced the template keep their slides and fall
        back to a built-in design on export. Built-in designs cannot be deleted.
      tags: [Templates]
      x-ptium-api-key-scopes: [templates:write]
      responses:
        '204':
          description: Template deleted.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/templates/{id}/download:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
      - $ref: '#/components/parameters/TemplateID'
    get:
      operationId: downloadTemplate
      summary: Download the stored template package
      tags: [Templates]
      x-ptium-api-key-scopes: [templates:read]
      responses:
        '200':
          description: The original PowerPoint package, byte for byte.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
            Content-Disposition:
              description: Attachment filename derived from the stored template.
              schema:
                type: string
          content:
            application/vnd.openxmlformats-officedocument.presentationml.presentation:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OIDC JWT or ptium_* API key
      description: |
        Send `Authorization: Bearer <credential>`. The credential may be an
        OIDC access-token JWT validated through issuer discovery and rotating
        JWKS, or an opaque Ptium API key beginning with `ptium_`.
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Alternative header for a Ptium `ptium_*` API key. Do not send it
        together with Authorization. The key must grant the scope named by the
        operation's `x-ptium-api-key-scopes` extension.
    DevSecret:
      type: apiKey
      in: header
      name: X-Ptium-Dev-Secret
      description: |
        Development-only shared secret. It is disabled by default, normally
        restricted to loopback clients, and may be renamed by server
        configuration. A Bearer value beginning `dev:` may accompany it to
        select an allowed development identity.
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Caller-generated correlation ID. Values longer than 128 characters are replaced.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of records. Values above 100 are clamped to 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    PresentationID:
      name: id
      in: path
      required: true
      description: Presentation UUID.
      schema:
        type: string
        format: uuid
...
  headers:
    XRequestID:
      description: Correlation ID used by logs, errors, and administrator incidents.
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    BadRequest:
      description: Malformed, oversized, trailing, or multi-value JSON; unknown fields; an invalid UUID path ID; or invalid transport input.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            invalidJSON:
              value:
                error:
                  code: invalid_json
                  message: Request body is not valid for this operation
                  details:
                    reason: 'json: unknown field "unexpected"'
                requestId: 8dcb8033f7d1f8027f2df88d
            invalidUUID:
              value:
                error:
                  code: invalid_id
                  message: Path id must be a UUID
                requestId: 8dcb8033f7d1f8027f2df88d
    Unauthorized:
      description: A credential is absent, invalid, expired, or revoked.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        WWW-Authenticate:
          description: Bearer authentication challenge.
          schema:
            type: string
            example: Bearer realm="ptium"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: authentication_required
              message: Unauthorized
            requestId: 8dcb8033f7d1f8027f2df88d
    Forbidden:
      description: The account is disabled or the API key lacks the operation's required scope.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            insufficientScope:
              value:
                error:
                  code: insufficient_scope
                  message: The API key does not grant this operation
                  details:
                    required: presentations:write
                requestId: 8dcb8033f7d1f8027f2df88d
            disabled:
              value:
                error:
                  code: account_disabled
                  message: This account has been disabled
                requestId: 8dcb8033f7d1f8027f2df88d
...
    NotFound:
      description: The record does not exist or is not owned by the current user.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: not_found
              message: The requested resource was not found
            requestId: 8dcb8033f7d1f8027f2df88d
...
  schemas:
...
    PresentationEnvelope:
      type: object
      additionalProperties: false
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/Presentation'
        requestId:
          $ref: '#/components/schemas/RequestID'
...